chore: add flash middleware
The flash middleware retrieves flash data from the user's session and adds it to the response locals before deleting it from the session. This data can be used to display messages or to pre-fill form data. It prevents the user from submitting the same form twice when refreshing on a post method by redirecting them after the initial submission and clearing the flash data from the session.